:recycle: Move test_gen_bids_sublist from bids_utils to test_bids_utils
